Refactor Dashboard component and introduce hooks for improved data management and functionality
- Streamlined the Dashboard component by removing redundant code and integrating new hooks: useDashboardData, useWorkflowActions, and useChatService.
- Enhanced data fetching and state management for workflows and chat functionality, ensuring a more modular and maintainable codebase.
- Preserved existing features while improving the overall structure and readability of the component.
